+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v0.98](https://github.com/realityforge/guiceyloops/tree/v0.98) (2019-02-15)
 [Full Changelog](https://github.com/realityforge/guiceyloops/compare/v0.96...v0.98)