+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Remove `{@inheritDoc}` as it only explicitly indicates that the default behaviour at the expense of significant visual clutter.
+
 ### [v0.98](https://github.com/realityforge/guiceyloops/tree/v0.98) (2019-02-15)
 [Full Changelog](https://github.com/realityforge/guiceyloops/compare/v0.96...v0.98)